report_parser test base class, expected_report_data
