+++ README.rst
@@ -75,6 +75,36 @@ templates coming soon.
 * `ScandiDos <http://scandidos.com>`__: *Delta4*
 
 
+Credits
+-------
+
+----------------
+Development Lead
+----------------
+
+* Dan Cutright
+
+------------
+Contributors
+------------
+
+* Marc Chamberland
+* Aditya Panchal
+
+
+---------
+Test Data
+---------
+Example IMRT QA reports used for unit testing and design are available `here <https://github.com/IQDM/IQDM-PDF/tree/master/tests/test_data/example_reports>`__.
+
+* Dan Cutright, University of Chicago Hospital
+    * delta4/UChicago
+    * sncpatient/UChicago
+
+* Serpil Kucuker Dogan, Nortwestern Memorial Hospital
+    * sncpatient/Northwestern_Memorial
+
+
 .. |build| image:: https://github.com/IQDM/IQDM-PDF/workflows/build/badge.svg
    :target: https://github.com/IQDM/IQDM-PDF/actions
    :alt: build
