+++ docs/source/authors.rst
@@ -12,3 +12,15 @@ Contributors
 
 * Marc Chamberland
 * Aditya Panchal
+
+
+Test Data
+---------
+Example IMRT QA reports used for unit testing and design are available `here <https://github.com/IQDM/IQDM-PDF/tree/master/tests/test_data/example_reports>`__.
+
+* Dan Cutright, University of Chicago Hospital
+    * delta4/UChicago
+    * sncpatient/UChicago
+
+* Serpil Kucuker Dogan, Nortwestern Memorial Hospital
+    * sncpatient/Northwestern_Memorial
